Create README.md
Uses a main class called BookGroup which acts as the center of the library. This class contains a reference to shelf, and shelf contains a refrence to books, and books contain a reference to a class called bookhistory. The BookGroup class has functions that help to access any part of the heirarchy. A person class at the top has information about them including their name and checkout history as well as a booklimit. The checkout process is automated in the program and every person checks out two books minimum. This program simply outputs book data and asks the user questions about books or the people that have checked them out.